refactor NoteForm jsx using react-hook-form, extract api calls into forlder
