+++ package.json
@@ -11,15 +11,18 @@
         "lint:fix": "eslint --fix ."
     },
     "dependencies": {
+        "@hookform/resolvers": "^3.1.0",
         "axios": "^1.2.3",
         "classnames": "^2.3.2",
         "react": "^18.2.0",
         "react-datepicker": "^4.10.0",
         "react-dom": "^18.2.0",
+        "react-hook-form": "^7.43.9",
         "react-modal": "^3.16.1",
         "react-router-dom": "^6.6.1",
         "sass": "^1.57.1",
-        "uuid": "^9.0.0"
+        "uuid": "^9.0.0",
+        "zod": "^3.21.4"
     },
     "devDependencies": {
         "@types/react": "^18.0.24",
