+++ src/components/Input/Input.tsx
@@ -1,5 +1,5 @@
 import classNames from "classnames";
-import { ChangeEventHandler } from "react";
+import { forwardRef } from "react";
 import { useTheme } from "@/context/ThemeContext";
 import styles from "./styles.module.scss";
 
@@ -8,39 +8,39 @@ type InputProps = {
     type: string;
     name?: string;
     placeholder?: string;
-    value: string | number;
-    onChange: ChangeEventHandler<HTMLInputElement>;
     disabled?: boolean;
     className?: string;
     id?: string;
 };
 
-function Input({
-    autoFocus = false,
-    disabled = false,
-    type,
-    name = "",
-    placeholder = "",
-    value,
-    onChange,
-    className = "",
-    id = "",
-}: InputProps) {
-    const { theme } = useTheme();
+const Input = forwardRef<HTMLInputElement, InputProps>(
+    (
+        {
+            autoFocus = false,
+            disabled = false,
+            type,
+            name = "",
+            placeholder = "",
+            className = "",
+            id = "",
+        },
+        ref
+    ) => {
+        const { theme } = useTheme();
 
-    const inputClass = classNames(styles.input, styles[theme], className);
-    return (
-        <input
-            autoFocus={autoFocus}
-            type={type}
-            name={name}
-            placeholder={placeholder}
-            value={value}
-            onChange={onChange}
-            className={inputClass}
-            id={id}
-            disabled={disabled}
-        />
-    );
-}
+        const inputClass = classNames(styles.input, styles[theme], className);
+        return (
+            <input
+                autoFocus={autoFocus}
+                type={type}
+                name={name}
+                placeholder={placeholder}
+                className={inputClass}
+                id={id}
+                disabled={disabled}
+                ref={ref}
+            />
+        );
+    }
+);
 export default Input;
