--- src/components/NoteForm/components/NoteFormActions/styles.module.scss
@@ -1,8 +1,0 @@
-@import "../../../../assets/variables";
-
-.actions {
-    display: flex;
-    width: 100%;
-    margin-top: 2em;
-    justify-content: space-between;
-}
