--- src/components/NoteForm/components/NoteInfo/styles.module.scss
@@ -1,15 +1,0 @@
-@import "../../../../assets/variables";
-
-.note_info {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    gap: 2em;
-    margin-bottom: 2em;
-}
-
-@media only screen and (max-width: $desktop-breakpoint) {
-    .input {
-        width: 100%;
-    }
-}
